Comply with MySQLdb >= 1.2.5
In MySQLdb 1.2.5, the parameter substitution
feature of execute() was generalized so that the second
parameter is required to be a tuple of values. Even if
there is only value to substitute, it still needs to
be passed as a tuple (i.e. execute(s, (x,)) rather than
just execute(s, x) when x of a simple (non-tuple) type).